refactor: decompose derive.rs into join.rs + classify.rs (#233)
Phase 3a: split derive.rs (1955 lines) into three focused modules:
- join.rs: worktree→PR→session matching and enrichment
- classify.rs: display group classification from joined data
- derive.rs: types, constants, phase/staleness helpers (351 lines)

Re-exports in derive.rs preserve backward compat for all callers.
