+++ crates/orchard/src/lib.rs
@@ -12,6 +12,7 @@ pub mod cache_sources;
 pub mod chat;
 pub mod ci_state;
 pub mod claude_state;
+pub mod classify;
 pub mod config;
 pub mod derive;
 pub mod events;
@@ -20,6 +21,7 @@ pub mod github;
 pub mod global_config;
 pub mod heal;
 pub mod hook_enrich;
+pub mod join;
 pub mod json_output;
 pub mod logger;
 pub mod models;
